More linter fixes
